requirements.txt
# Зависимости проекта
numpy>=1.19.5
pillow>=8.0.0
tensorflow>=2.5.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tqdm>=4.60.0
pandas>=1.2.0
opencv-python>=4.5.0